fix: update stubs and `IIIFFile`
